Refactor project to make program sections clearer.
