+++ src/MusicSharp.csproj
@@ -3,7 +3,7 @@
   <PropertyGroup>
     <OutputType>Exe</OutputType>
     <TargetFramework>net8.0</TargetFramework>
-    <Version>0.8.9</Version>
+    <Version>1.0.0</Version>
     <Authors>Mark-James McDougall</Authors>
     <Company>Mark-James McDougall</Company>
     <Nullable>enable</Nullable>
